pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "name-of-the-machine"
version = "0.0.1"
authors = [
    {name="Alexander Fedotov", email="alex.fedotov@aol.com"},
]
description = "Description of the Machine"
readme = "README.md"
requires-python = ">=3.10"
classifiers=[
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
]
dependencies = [
    "PyGithub >= 2.3.0",
    "PyYAML >= 6.0.1"
]
[project.urls]
"Homepage" = "https://github.com/org_name/package_name"
"Bug Tracker" = "https://github.com/org_name/package_name/issues"